pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "holdings-monitor"
version = "0.2.0"
description = "Config-driven holdings disclosure monitor with snapshotting, diffing, validation, and notifications."
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
  {name = "OpenAI generated starter for public side-project hardening"}
]
dependencies = [
  "pandas>=2.2.0",
  "openpyxl>=3.1.0",
  "requests>=2.32.0",
  "PyYAML>=6.0.0",
  "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "ruff>=0.6.0",
]

[project.scripts]
holdings-monitor = "holdings_monitor.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["B008"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
